Refactor backup/restore logic and implement BackupRestoreViewNewView
- Centralized backup and restore logic into `BackupRestoreManager.swift`.
- Moved shared UI components to `BackupRestoreUIComponents.swift`.
- Updated existing `BackupRestoreView.swift` to use the new manager.
- Created `BackupRestoreViewNewView.swift` with a simplified, focused UI and fully implemented backup/import functionality.
- Fixed several bugs in the backup/restore flow and improved it with async/await.
