+++ Feather/Views/Settings/Backup & Restore/BackupRestoreView.swift
@@ -6,30 +6,19 @@ import ZIPFoundation
 import UniformTypeIdentifiers
 import CoreData
 
-// MARK: - Backup Options
-struct BackupOptions {
-	var includeCertificates: Bool = true
-	var includeSignedApps: Bool = true
-	var includeImportedApps: Bool = true
-	var includeSources: Bool = true
-	var includeDefaultFrameworks: Bool = true
-}
-
 // MARK: - View
 struct BackupRestoreView: View {
 	@Environment(\.dismiss) var dismiss
+	@StateObject private var manager = BackupRestoreManager.shared
 	@State private var isRestoreFilePickerPresented = false
 	@State private var isVerifyFilePickerPresented = false
 	@State private var showRestoreDialog = false
 	@State private var pendingRestoreURL: URL?
 	@State private var isBackupOptionsPresented = false
 	@State private var backupOptions = BackupOptions()
-	@State private var isRestoring = false
-	@State private var restoreProgress: Double = 0.0
 	@State private var showInvalidBackupError = false
 	@State private var backupDocument: BackupDocument?
 	@State private var showExporter = false
-	@State private var isVerifying = false
 	@AppStorage("feature_advancedBackupTools") var advancedBackupTools = false
 	
 	// MARK: Body
@@ -208,7 +197,15 @@ struct BackupRestoreView: View {
 			if advancedBackupTools {
 				Section {
 					Button {
-						exportFullDatabase()
+						Task {
+							do {
+								let url = try await manager.exportFullDatabase()
+								backupDocument = BackupDocument(url: url)
+								showExporter = true
+							} catch {
+								UIAlertController.showAlertWithOk(title: "Error", message: error.localizedDescription)
+							}
+						}
 					} label: {
 						Label(.localized("Export Full Database"), systemImage: "cylinder.split.1x2.fill")
 					}
@@ -247,9 +244,14 @@ struct BackupRestoreView: View {
 				options: $backupOptions,
 				onConfirm: {
 					isBackupOptionsPresented = false
-					if let url = prepareBackup(with: backupOptions) {
-						backupDocument = BackupDocument(url: url)
-						showExporter = true
+					Task {
+						do {
+							let url = try await manager.prepareBackup(with: backupOptions)
+							backupDocument = BackupDocument(url: url)
+							showExporter = true
+						} catch {
+							UIAlertController.showAlertWithOk(title: .localized("Error"), message: error.localizedDescription)
+						}
 					}
 				}
 			)
@@ -274,7 +276,18 @@ struct BackupRestoreView: View {
 				onDocumentsPicked: { urls in
 					isVerifyFilePickerPresented = false
 					guard let url = urls.first else { return }
-					verifyBackup(at: url)
+					Task {
+						do {
+							let isValid = try await manager.verifyBackup(at: url)
+							if isValid {
+								UIAlertController.showAlertWithOk(title: .localized("Verification Successful"), message: .localized("This backup file is valid and can be restored."))
+							} else {
+								showInvalidBackupError = true
+							}
+						} catch {
+							UIAlertController.showAlertWithOk(title: .localized("Error"), message: .localized("Failed to verify backup: \(error.localizedDescription)"))
+						}
+					}
 				}
 			)
 			.ignoresSafeArea()
@@ -309,7 +322,13 @@ struct BackupRestoreView: View {
 			}
 			Button(.localized("Yes")) {
 				if let url = pendingRestoreURL {
-					performRestore(from: url, restart: true)
+					Task {
+						do {
+							try await manager.performRestore(from: url, restart: true)
+						} catch {
+							UIAlertController.showAlertWithOk(title: .localized("Error"), message: error.localizedDescription)
+						}
+					}
 				}
 			}
 		} message: {
@@ -321,11 +340,11 @@ struct BackupRestoreView: View {
 			Text(.localized("Not a valid Backup file because Portal couldn't find the checker inside the file. Please upload an actual .zip file of a backup."))
 		}
 		.overlay {
-			if isRestoring {
-				RestoreLoadingOverlay(progress: restoreProgress)
+			if manager.isRestoring {
+				RestoreLoadingOverlay(progress: manager.restoreProgress)
 			}
 
-			if isVerifying {
+			if manager.isVerifying {
 				ZStack {
 					Color.black.opacity(0.4).ignoresSafeArea()
 					VStack(spacing: 12) {
@@ -365,984 +384,5 @@ struct BackupRestoreView: View {
 		.padding(.vertical, 4)
 	}
 	
-	// MARK: - Advanced Tools Functions
-	private func verifyBackup(at url: URL) {
-		isVerifying = true
-		let tempDir = FileManager.default.temporaryDirectory.appendingPathComponent(UUID().uuidString)
-
-		do {
-			// FileImporterRepresentableView uses asCopy: true, so the file is already copied
-			// and we don't need security-scoped resource access. Calling it would return false.
-			try FileManager.default.createDirectory(at: tempDir, withIntermediateDirectories: true)
-			try FileManager.default.unzipItem(at: url, to: tempDir)
-
-			let markers = ["PORTAL_BACKUP_MARKER.txt", "FEATHER_BACKUP_MARKER.txt", "PORTAL_BACKUP_CHECKER.txt"]
-			let hasMarker = markers.contains { marker in
-				FileManager.default.fileExists(atPath: tempDir.appendingPathComponent(marker).path)
-			}
-
-			let hasSettings = FileManager.default.fileExists(atPath: tempDir.appendingPathComponent("settings.plist").path)
-
-			try? FileManager.default.removeItem(at: tempDir)
-			isVerifying = false
-
-			if hasMarker && hasSettings {
-				UIAlertController.showAlertWithOk(title: .localized("Verification Successful"), message: .localized("This backup file is valid and can be restored."))
-			} else {
-				showInvalidBackupError = true
-			}
-		} catch {
-			isVerifying = false
-			UIAlertController.showAlertWithOk(title: .localized("Error"), message: .localized("Failed to verify backup: \(error.localizedDescription)"))
-		}
-	}
-
-	private func exportFullDatabase() {
-		guard let storeURL = Storage.shared.container.persistentStoreDescriptions.first?.url else {
-			UIAlertController.showAlertWithOk(title: "Error", message: "Could not find database location")
-			return
-		}
-
-		let tempDir = FileManager.default.temporaryDirectory.appendingPathComponent(UUID().uuidString)
-		let zipURL = FileManager.default.temporaryDirectory.appendingPathComponent("PortalDatabaseBackup_\(Date().timeIntervalSince1970).zip")
-
-		do {
-			try FileManager.default.createDirectory(at: tempDir, withIntermediateDirectories: true)
-
-			// Copy SQLite files (including SHM and WAL)
-			let baseName = storeURL.lastPathComponent
-			let directory = storeURL.deletingLastPathComponent()
-
-			let filesToCopy = [baseName, "\(baseName)-shm", "\(baseName)-wal"]
-			for fileName in filesToCopy {
-				let fileURL = directory.appendingPathComponent(fileName)
-				if FileManager.default.fileExists(atPath: fileURL.path) {
-					try FileManager.default.copyItem(at: fileURL, to: tempDir.appendingPathComponent(fileName))
-				}
-			}
-
-			try FileManager.default.zipItem(at: tempDir, to: zipURL, shouldKeepParent: false)
-			try? FileManager.default.removeItem(at: tempDir)
-
-			backupDocument = BackupDocument(url: zipURL)
-			showExporter = true
-
-		} catch {
-			UIAlertController.showAlertWithOk(title: "Error", message: "Failed to export database: \(error.localizedDescription)")
-		}
-	}
-
-	// MARK: - Backup Functions
-	private func prepareBackup(with options: BackupOptions) -> URL? {
-		// Create temporary directory for backup
-		let tempDir = FileManager.default.temporaryDirectory.appendingPathComponent(UUID().uuidString)
-		
-		do {
-			try FileManager.default.createDirectory(at: tempDir, withIntermediateDirectories: true)
-			
-			// 1. Backup certificates with full metadata (if selected)
-			if options.includeCertificates {
-				let certificatesDir = tempDir.appendingPathComponent("certificates")
-				try? FileManager.default.createDirectory(at: certificatesDir, withIntermediateDirectories: true)
-				let certificates = Storage.shared.getAllCertificates()
-				var certMetadata: [[String: Any]] = []
-				
-				for cert in certificates {
-					if let uuid = cert.uuid {
-						var metadata: [String: Any] = ["uuid": uuid]
-						
-						// Save certificate file
-						if let certURL = Storage.shared.getFile(.certificate, from: cert),
-						   let certData = try? Data(contentsOf: certURL) {
-							let destURL = certificatesDir.appendingPathComponent("\(uuid).p12")
-							try certData.write(to: destURL)
-							metadata["hasP12"] = true
-						}
-						
-						// Save provisioning profile
-						if let provisionURL = Storage.shared.getFile(.provision, from: cert),
-						   let provisionData = try? Data(contentsOf: provisionURL) {
-							let destURL = certificatesDir.appendingPathComponent("\(uuid).mobileprovision")
-							try provisionData.write(to: destURL)
-							metadata["hasProvision"] = true
-						}
-						
-						// Save metadata
-						if let provisionData = Storage.shared.getProvisionFileDecoded(for: cert) {
-							metadata["name"] = provisionData.Name
-							if let teamID = provisionData.TeamIdentifier.first {
-								metadata["teamID"] = teamID
-							}
-							metadata["teamName"] = provisionData.TeamName
-						}
-						if let date = cert.date { metadata["date"] = date.timeIntervalSince1970 }
-						metadata["ppQCheck"] = cert.ppQCheck
-						if let password = cert.password { metadata["password"] = password }
-						
-						certMetadata.append(metadata)
-					}
-				}
-				
-				// Save certificate metadata
-				let certMetadataFile = tempDir.appendingPathComponent("certificates_metadata.json")
-				if let jsonData = try? JSONSerialization.data(withJSONObject: certMetadata) {
-					try jsonData.write(to: certMetadataFile)
-				}
-			}
-			
-			// 2. Backup sources (if selected)
-			if options.includeSources {
-				let sourcesFile = tempDir.appendingPathComponent("sources.json")
-				let sources = Storage.shared.getSources()
-				let sourcesData = sources.compactMap { source -> [String: String]? in
-					guard let urlString = source.sourceURL?.absoluteString,
-						  let name = source.name,
-						  let identifier = source.identifier else { return nil }
-					return ["url": urlString, "name": name, "identifier": identifier]
-				}
-				if let jsonData = try? JSONSerialization.data(withJSONObject: sourcesData) {
-					try jsonData.write(to: sourcesFile)
-				}
-			}
-			
-			// 3. Backup signed apps (if selected)
-			if options.includeSignedApps {
-				let signedAppsFile = tempDir.appendingPathComponent("signed_apps.json")
-				let signedAppsDir = tempDir.appendingPathComponent("signed_apps")
-				try? FileManager.default.createDirectory(at: signedAppsDir, withIntermediateDirectories: true)
-				
-				let fetchRequest = Signed.fetchRequest()
-				if let signedApps = try? Storage.shared.context.fetch(fetchRequest) {
-					var appsData: [[String: String]] = []
-					
-					for app in signedApps {
-						guard let uuid = app.uuid else { continue }
-						var data: [String: String] = ["uuid": uuid]
-						if let name = app.name { data["name"] = name }
-						if let identifier = app.identifier { data["identifier"] = identifier }
-						if let version = app.version { data["version"] = version }
-						
-						// Copy actual IPA file if it exists
-						let signedDir = FileManager.default.signed(uuid)
-						if let ipaURL = FileManager.default.getPath(in: signedDir, for: "ipa"),
-						   FileManager.default.fileExists(atPath: ipaURL.path) {
-							let destURL = signedAppsDir.appendingPathComponent("\(uuid).ipa")
-							try? FileManager.default.copyItem(at: ipaURL, to: destURL)
-							data["hasIPA"] = "true"
-						}
-						
-						appsData.append(data)
-					}
-					
-					if let jsonData = try? JSONSerialization.data(withJSONObject: appsData) {
-						try jsonData.write(to: signedAppsFile)
-					}
-				}
-			}
-			
-			// 4. Backup imported apps (if selected)
-			if options.includeImportedApps {
-				let importedAppsFile = tempDir.appendingPathComponent("imported_apps.json")
-				let importedAppsDir = tempDir.appendingPathComponent("imported_apps")
-				try? FileManager.default.createDirectory(at: importedAppsDir, withIntermediateDirectories: true)
-				
-				let fetchRequest = Imported.fetchRequest()
-				if let importedApps = try? Storage.shared.context.fetch(fetchRequest) {
-					var appsData: [[String: String]] = []
-					
-					for app in importedApps {
-						guard let uuid = app.uuid else { continue }
-						var data: [String: String] = ["uuid": uuid]
-						if let name = app.name { data["name"] = name }
-						if let identifier = app.identifier { data["identifier"] = identifier }
-						if let version = app.version { data["version"] = version }
-						
-						// Copy actual IPA file if it exists
-						let importedDir = FileManager.default.unsigned(uuid)
-						if let ipaURL = FileManager.default.getPath(in: importedDir, for: "ipa"),
-						   FileManager.default.fileExists(atPath: ipaURL.path) {
-							let destURL = importedAppsDir.appendingPathComponent("\(uuid).ipa")
-							try? FileManager.default.copyItem(at: ipaURL, to: destURL)
-							data["hasIPA"] = "true"
-						}
-						
-						appsData.append(data)
-					}
-					
-					if let jsonData = try? JSONSerialization.data(withJSONObject: appsData) {
-						try jsonData.write(to: importedAppsFile)
-					}
-				}
-			}
-			
-			// 4.5 Backup default frameworks (if selected)
-			if options.includeDefaultFrameworks {
-				let defaultFrameworksDir = tempDir.appendingPathComponent("default_frameworks")
-				try? FileManager.default.createDirectory(at: defaultFrameworksDir, withIntermediateDirectories: true)
-
-				let sourceDFDir = Storage.shared.documentsURL.appendingPathComponent("Feather/DefaultFrameworks")
-				if FileManager.default.fileExists(atPath: sourceDFDir.path) {
-					let contents = try? FileManager.default.contentsOfDirectory(at: sourceDFDir, includingPropertiesForKeys: nil)
-					for fileURL in contents ?? [] {
-						try? FileManager.default.copyItem(at: fileURL, to: defaultFrameworksDir.appendingPathComponent(fileURL.lastPathComponent))
-					}
-				}
-			}
-
-			// 4.6 Backup archives (if selected)
-			if options.includeSignedApps {
-				let archivesDir = tempDir.appendingPathComponent("archives")
-				try? FileManager.default.createDirectory(at: archivesDir, withIntermediateDirectories: true)
-
-				let sourceArchivesDir = FileManager.default.archives
-				if FileManager.default.fileExists(atPath: sourceArchivesDir.path) {
-					let contents = try? FileManager.default.contentsOfDirectory(at: sourceArchivesDir, includingPropertiesForKeys: nil)
-					for fileURL in contents ?? [] {
-						try? FileManager.default.copyItem(at: fileURL, to: archivesDir.appendingPathComponent(fileURL.lastPathComponent))
-					}
-				}
-			}
-
-			// 4.7 Backup pairing and SSL files (always included as part of state)
-			let filesToBackup = ["pairingFile.plist", "server.pem", "server.crt", "commonName.txt"]
-			let extraFilesDir = tempDir.appendingPathComponent("extra_files")
-			try? FileManager.default.createDirectory(at: extraFilesDir, withIntermediateDirectories: true)
-
-			for fileName in filesToBackup {
-				let fileURL = Storage.shared.documentsURL.appendingPathComponent(fileName)
-				if FileManager.default.fileExists(atPath: fileURL.path) {
-					try? FileManager.default.copyItem(at: fileURL, to: extraFilesDir.appendingPathComponent(fileName))
-				}
-			}
-
-			// 4.8 Backup Core Data database
-			let databaseDir = tempDir.appendingPathComponent("database")
-			try? FileManager.default.createDirectory(at: databaseDir, withIntermediateDirectories: true)
-
-			if let storeURL = Storage.shared.container.persistentStoreDescriptions.first?.url {
-				let baseName = storeURL.lastPathComponent
-				let directory = storeURL.deletingLastPathComponent()
-				let dbFiles = [baseName, "\(baseName)-shm", "\(baseName)-wal"]
-
-				for dbFile in dbFiles {
-					let fileURL = directory.appendingPathComponent(dbFile)
-					if FileManager.default.fileExists(atPath: fileURL.path) {
-						try? FileManager.default.copyItem(at: fileURL, to: databaseDir.appendingPathComponent(dbFile))
-					}
-				}
-			}
-
-			// 5. Backup ALL settings - always included
-			let settingsFile = tempDir.appendingPathComponent("settings.plist")
-			let defaults = UserDefaults.standard.dictionaryRepresentation()
-			// Include all Feather and app-specific settings
-			let filtered = defaults.filter { key, _ in
-				!key.hasPrefix("NS") &&
-				!key.hasPrefix("AK") &&
-				!key.hasPrefix("Apple") &&
-				!key.hasPrefix("WebKit") &&
-				!key.hasPrefix("CPU") &&
-				!key.hasPrefix("metal")
-			}
-			let settingsData = try PropertyListSerialization.data(fromPropertyList: filtered, format: .xml, options: 0)
-			try settingsData.write(to: settingsFile)
-			
-			// 6. Create zip file with validation marker
-			let tempBackupDir = FileManager.default.temporaryDirectory.appendingPathComponent("Backups")
-			try? FileManager.default.createDirectory(at: tempBackupDir, withIntermediateDirectories: true)
-
-			let backupFileName = "PortalBackup_\(UUID().uuidString).zip"
-			let finalZipURL = tempBackupDir.appendingPathComponent(backupFileName)
-			
-			// Remove existing file if present
-			try? FileManager.default.removeItem(at: finalZipURL)
-			
-			// Add a backup marker file to validate later
-			let markerFile = tempDir.appendingPathComponent("PORTAL_BACKUP_MARKER.txt")
-			let markerContent = "PORTAL_BACKUP_v1.0_\(Date().timeIntervalSince1970)"
-			try markerContent.write(to: markerFile, atomically: true, encoding: .utf8)
-			
-			try FileManager.default.zipItem(at: tempDir, to: finalZipURL, shouldKeepParent: false)
-			
-			// Clean up temp directory
-			try? FileManager.default.removeItem(at: tempDir)
-			
-			return finalZipURL
-			
-		} catch {
-			UIAlertController.showAlertWithOk(
-				title: .localized("Error"),
-				message: .localized("Failed to prepare backup: \(error.localizedDescription)")
-			)
-			return nil
-		}
-	}
-	
-	private func performRestore(from url: URL, restart: Bool) {
-		let tempDir = FileManager.default.temporaryDirectory.appendingPathComponent(UUID().uuidString)
-		
-		// Start restore animation
-		isRestoring = true
-		restoreProgress = 0.0
-		
-		do {
-			// FileImporterRepresentableView uses asCopy: true, so the file is already copied
-			// and we don't need security-scoped resource access. Calling it would return false.
-			try FileManager.default.createDirectory(at: tempDir, withIntermediateDirectories: true)
-			
-			// Unzip backup
-			withAnimation {
-				restoreProgress = 0.1
-			}
-			try FileManager.default.unzipItem(at: url, to: tempDir)
-			
-			withAnimation {
-				restoreProgress = 0.2
-			}
-			
-			// VALIDATE BACKUP: Check for marker file (including legacy ones)
-			let markers = ["PORTAL_BACKUP_MARKER.txt", "FEATHER_BACKUP_MARKER.txt", "PORTAL_BACKUP_CHECKER.txt"]
-			let hasMarker = markers.contains { marker in
-				let path = tempDir.appendingPathComponent(marker).path
-				if FileManager.default.fileExists(atPath: path) {
-					if let content = try? String(contentsOfFile: path, encoding: .utf8),
-					   (content.contains("PORTAL_BACKUP") || content.contains("FEATHER_BACKUP")) {
-						return true
-					}
-				}
-				return false
-			}
-
-			guard hasMarker else {
-				// Clean up
-				try? FileManager.default.removeItem(at: tempDir)
-				isRestoring = false
-				showInvalidBackupError = true
-				return
-			}
-			
-			// Additional validation: check for at least one expected file
-			let settingsFile = tempDir.appendingPathComponent("settings.plist")
-			let hasValidContent = FileManager.default.fileExists(atPath: settingsFile.path)
-			
-			if !hasValidContent {
-				// Clean up
-				try? FileManager.default.removeItem(at: tempDir)
-				isRestoring = false
-				showInvalidBackupError = true
-				return
-			}
-			
-			withAnimation {
-				restoreProgress = 0.3
-			}
-			
-			// 1. Restore certificates with metadata
-			let certificatesDir = tempDir.appendingPathComponent("certificates")
-			let certMetadataFile = tempDir.appendingPathComponent("certificates_metadata.json")
-			
-			if FileManager.default.fileExists(atPath: certMetadataFile.path),
-			   let jsonData = try? Data(contentsOf: certMetadataFile),
-			   let metadata = try? JSONSerialization.jsonObject(with: jsonData) as? [[String: Any]] {
-				
-				for certInfo in metadata {
-					guard let uuid = certInfo["uuid"] as? String else { continue }
-					
-					let p12URL = certificatesDir.appendingPathComponent("\(uuid).p12")
-					let provisionURL = certificatesDir.appendingPathComponent("\(uuid).mobileprovision")
-					
-					// Only restore if both files exist
-					if FileManager.default.fileExists(atPath: p12URL.path),
-					   FileManager.default.fileExists(atPath: provisionURL.path) {
-						
-						// Copy certificates to the proper storage location
-						let certStorageURL = Storage.shared.documentsURL.appendingPathComponent("certificates/\(uuid).p12")
-						let provisionStorageURL = Storage.shared.documentsURL.appendingPathComponent("certificates/\(uuid).mobileprovision")
-						
-						// Ensure certificates directory exists
-						try? FileManager.default.createDirectory(
-							at: Storage.shared.documentsURL.appendingPathComponent("certificates"),
-							withIntermediateDirectories: true
-						)
-						
-						try? FileManager.default.copyItem(at: p12URL, to: certStorageURL)
-						try? FileManager.default.copyItem(at: provisionURL, to: provisionStorageURL)
-						
-						let name = certInfo["name"] as? String ?? "Restored Certificate"
-						Storage.shared.addCertificate(
-							uuid: uuid,
-							password: certInfo["password"] as? String,
-							nickname: name,
-							ppq: certInfo["ppQCheck"] as? Bool ?? false,
-							expiration: Date(timeIntervalSince1970: certInfo["date"] as? Double ?? Date().timeIntervalSince1970),
-							completion: { _ in }
-						)
-						AppLogManager.shared.info("Restored certificate: \(name)", category: "Backup & Restore")
-					}
-				}
-			}
-			
-			withAnimation {
-				restoreProgress = 0.5
-			}
-			
-			// 2. Restore sources
-			let sourcesFile = tempDir.appendingPathComponent("sources.json")
-			if FileManager.default.fileExists(atPath: sourcesFile.path) {
-				let jsonData = try Data(contentsOf: sourcesFile)
-				if let sources = try JSONSerialization.jsonObject(with: jsonData) as? [[String: String]] {
-					for source in sources {
-						if let urlString = source["url"], 
-						   let sourceURL = URL(string: urlString),
-						   let name = source["name"] {
-							let identifier = source["identifier"] ?? sourceURL.absoluteString
-							Storage.shared.addSource(
-								sourceURL,
-								name: name,
-								identifier: identifier,
-								completion: { _ in }
-							)
-						}
-					}
-				}
-			}
-			
-			withAnimation {
-				restoreProgress = 0.65
-			}
-			
-			// 3. Restore signed apps with actual IPA files
-			let signedAppsFile = tempDir.appendingPathComponent("signed_apps.json")
-			let signedAppsDir = tempDir.appendingPathComponent("signed_apps")
-			if FileManager.default.fileExists(atPath: signedAppsFile.path) {
-				let jsonData = try Data(contentsOf: signedAppsFile)
-				if let appsData = try JSONSerialization.jsonObject(with: jsonData) as? [[String: String]] {
-					for appInfo in appsData {
-						guard let uuid = appInfo["uuid"],
-							  let hasIPA = appInfo["hasIPA"],
-							  hasIPA == "true" else { continue }
-						
-						let ipaSourceURL = signedAppsDir.appendingPathComponent("\(uuid).ipa")
-						if FileManager.default.fileExists(atPath: ipaSourceURL.path) {
-							// Ensure signed apps directory exists
-							let signedStorageDir = Storage.shared.documentsURL.appendingPathComponent("signed")
-							try? FileManager.default.createDirectory(at: signedStorageDir, withIntermediateDirectories: true)
-							
-							let ipaDestURL = signedStorageDir.appendingPathComponent("\(uuid).ipa")
-							try? FileManager.default.copyItem(at: ipaSourceURL, to: ipaDestURL)
-							
-							if let name = appInfo["name"] {
-								Storage.shared.addSigned(
-									uuid: uuid,
-									appName: name,
-									appIdentifier: appInfo["identifier"],
-									appVersion: appInfo["version"],
-									completion: { _ in }
-								)
-								AppLogManager.shared.info("Restored signed app: \(name)", category: "Backup & Restore")
-							}
-						}
-					}
-				}
-			}
-			
-			withAnimation {
-				restoreProgress = 0.8
-			}
-			
-			// 4. Restore imported apps with actual IPA files
-			let importedAppsFile = tempDir.appendingPathComponent("imported_apps.json")
-			let importedAppsDir = tempDir.appendingPathComponent("imported_apps")
-			if FileManager.default.fileExists(atPath: importedAppsFile.path) {
-				let jsonData = try Data(contentsOf: importedAppsFile)
-				if let appsData = try JSONSerialization.jsonObject(with: jsonData) as? [[String: String]] {
-					for appInfo in appsData {
-						guard let uuid = appInfo["uuid"],
-							  let hasIPA = appInfo["hasIPA"],
-							  hasIPA == "true" else { continue }
-						
-						let ipaSourceURL = importedAppsDir.appendingPathComponent("\(uuid).ipa")
-						if FileManager.default.fileExists(atPath: ipaSourceURL.path) {
-							// Ensure imported apps directory exists
-							let importedStorageDir = Storage.shared.documentsURL.appendingPathComponent("imported")
-							try? FileManager.default.createDirectory(at: importedStorageDir, withIntermediateDirectories: true)
-							
-							let ipaDestURL = importedStorageDir.appendingPathComponent("\(uuid).ipa")
-							try? FileManager.default.copyItem(at: ipaSourceURL, to: ipaDestURL)
-							
-							if let name = appInfo["name"] {
-								Storage.shared.addImported(
-									uuid: uuid,
-									appName: name,
-									appIdentifier: appInfo["identifier"],
-									appVersion: appInfo["version"],
-									completion: { _ in }
-								)
-								AppLogManager.shared.info("Restored imported app: \(name)", category: "Backup & Restore")
-							}
-						}
-					}
-				}
-			}
-			
-			// 4.5 Restore default frameworks
-			let defaultFrameworksDir = tempDir.appendingPathComponent("default_frameworks")
-			if FileManager.default.fileExists(atPath: defaultFrameworksDir.path) {
-				let destDFDir = Storage.shared.documentsURL.appendingPathComponent("Feather/DefaultFrameworks")
-				try? FileManager.default.createDirectory(at: destDFDir, withIntermediateDirectories: true)
-
-				let contents = try? FileManager.default.contentsOfDirectory(at: defaultFrameworksDir, includingPropertiesForKeys: nil)
-				for fileURL in contents ?? [] {
-					let destURL = destDFDir.appendingPathComponent(fileURL.lastPathComponent)
-					if FileManager.default.fileExists(atPath: destURL.path) {
-						try? FileManager.default.removeItem(at: destURL)
-					}
-					try? FileManager.default.copyItem(at: fileURL, to: destURL)
-				}
-				AppLogManager.shared.info("Restored default frameworks", category: "Backup & Restore")
-			}
-
-			// 4.6 Restore archives
-			let archivesDir = tempDir.appendingPathComponent("archives")
-			if FileManager.default.fileExists(atPath: archivesDir.path) {
-				let destArchivesDir = FileManager.default.archives
-				try? FileManager.default.createDirectory(at: destArchivesDir, withIntermediateDirectories: true)
-
-				let contents = try? FileManager.default.contentsOfDirectory(at: archivesDir, includingPropertiesForKeys: nil)
-				for fileURL in contents ?? [] {
-					let destURL = destArchivesDir.appendingPathComponent(fileURL.lastPathComponent)
-					if FileManager.default.fileExists(atPath: destURL.path) {
-						try? FileManager.default.removeItem(at: destURL)
-					}
-					try? FileManager.default.copyItem(at: fileURL, to: destURL)
-				}
-				AppLogManager.shared.info("Restored archives", category: "Backup & Restore")
-			}
-
-			// 4.7 Restore pairing and SSL files
-			let extraFilesDir = tempDir.appendingPathComponent("extra_files")
-			if FileManager.default.fileExists(atPath: extraFilesDir.path) {
-				let contents = try? FileManager.default.contentsOfDirectory(at: extraFilesDir, includingPropertiesForKeys: nil)
-				for fileURL in contents ?? [] {
-					let destURL = Storage.shared.documentsURL.appendingPathComponent(fileURL.lastPathComponent)
-					if FileManager.default.fileExists(atPath: destURL.path) {
-						try? FileManager.default.removeItem(at: destURL)
-					}
-					try? FileManager.default.copyItem(at: fileURL, to: destURL)
-				}
-				AppLogManager.shared.info("Restored extra state files", category: "Backup & Restore")
-			}
-
-			// 4.8 Restore database
-			let databaseDirInBackup = tempDir.appendingPathComponent("database")
-			if FileManager.default.fileExists(atPath: databaseDirInBackup.path),
-			   let storeURL = Storage.shared.container.persistentStoreDescriptions.first?.url {
-				let baseName = storeURL.lastPathComponent
-				let directory = storeURL.deletingLastPathComponent()
-				let dbFiles = [baseName, "\(baseName)-shm", "\(baseName)-wal"]
-
-				for dbFile in dbFiles {
-					let srcURL = databaseDirInBackup.appendingPathComponent(dbFile)
-					let destURL = directory.appendingPathComponent(dbFile)
-					if FileManager.default.fileExists(atPath: srcURL.path) {
-						if FileManager.default.fileExists(atPath: destURL.path) {
-							try? FileManager.default.removeItem(at: destURL)
-						}
-						try? FileManager.default.copyItem(at: srcURL, to: destURL)
-					}
-				}
-				AppLogManager.shared.info("Restored database files", category: "Backup & Restore")
-			}
-
-			withAnimation {
-				restoreProgress = 0.9
-			}
-			
-			// 5. Restore ALL settings
-			if FileManager.default.fileExists(atPath: settingsFile.path) {
-				let settingsData = try Data(contentsOf: settingsFile)
-				if let settings = try PropertyListSerialization.propertyList(from: settingsData, options: [], format: nil) as? [String: Any] {
-					for (key, value) in settings {
-						// Restore all settings except system-specific ones
-						if !key.hasPrefix("NS") && !key.hasPrefix("AK") && !key.hasPrefix("Apple") {
-							UserDefaults.standard.set(value, forKey: key)
-						}
-					}
-					UserDefaults.standard.synchronize()
-				}
-			}
-			
-			withAnimation {
-				restoreProgress = 1.0
-			}
-			
-			// Clean up
-			try? FileManager.default.removeItem(at: tempDir)
-			
-			// Small delay to show completion
-			DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
-				isRestoring = false
-				
-				if restart {
-					// Restart the app
-					UIAlertController.showAlertWithOk(
-						title: .localized("Restore Complete"),
-						message: .localized("The app will now restart to apply changes.")
-					) {
-						DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
-							UIApplication.shared.suspendAndReopen()
-						}
-					}
-				} else {
-					UIAlertController.showAlertWithOk(
-						title: .localized("Success"),
-						message: .localized("Backup restored successfully. Changes will be applied on next restart.")
-					)
-				}
-			}
-			
-		} catch {
-			isRestoring = false
-			UIAlertController.showAlertWithOk(
-				title: .localized("Error"),
-				message: .localized("Failed to restore backup: \(error.localizedDescription)")
-			)
-		}
-	}
-}
-
-// MARK: - BackupDocument
-struct BackupDocument: FileDocument {
-	static var readableContentTypes: [UTType] { [.zip] }
-	
-	var url: URL
-	
-	init(url: URL) {
-		self.url = url
-	}
-	
-	init(configuration: ReadConfiguration) throws {
-		// For reading, we don't need to handle this as we're only exporting
-		throw CocoaError(.fileReadUnknown)
-	}
-	
-	func fileWrapper(configuration: WriteConfiguration) throws -> FileWrapper {
-		return try FileWrapper(url: url)
-	}
-}
-
-// MARK: - BackupOptionsView
-struct BackupOptionsView: View {
-	@Environment(\.dismiss) var dismiss
-	@Binding var options: BackupOptions
-	let onConfirm: () -> Void
-	
-	var body: some View {
-		NavigationView {
-			ScrollView {
-				VStack(spacing: 24) {
-					// Header
-					VStack(spacing: 16) {
-						ZStack {
-							Circle()
-								.fill(
-									LinearGradient(
-										gradient: Gradient(colors: [Color.blue.opacity(0.2), Color.purple.opacity(0.2)]),
-										startPoint: .topLeading,
-										endPoint: .bottomTrailing
-									)
-								)
-								.frame(width: 80, height: 80)
-							
-							Image(systemName: "square.and.arrow.up.fill")
-								.font(.system(size: 40, weight: .semibold))
-								.foregroundStyle(
-									LinearGradient(
-										gradient: Gradient(colors: [.blue, .purple]),
-										startPoint: .topLeading,
-										endPoint: .bottomTrailing
-									)
-								)
-						}
-						
-						Text(.localized("What would you like in this Portal Backup?"))
-							.font(.title2.bold())
-							.multilineTextAlignment(.center)
-							.padding(.horizontal)
-						
-						Text(.localized("Select the data you want to include in your backup"))
-							.font(.subheadline)
-							.foregroundStyle(.secondary)
-							.multilineTextAlignment(.center)
-							.padding(.horizontal)
-					}
-					.padding(.top, 20)
-					
-					// Options
-					VStack(spacing: 12) {
-						backupOptionToggle(
-							icon: "checkmark.seal.fill",
-							iconColor: .blue,
-							title: .localized("Certificates"),
-							description: .localized("Your signing certificates and provisioning profiles"),
-							isOn: $options.includeCertificates
-						)
-						
-						backupOptionToggle(
-							icon: "app.badge.fill",
-							iconColor: .green,
-							title: .localized("Signed Apps"),
-							description: .localized("Apps you have signed with your certificates"),
-							isOn: $options.includeSignedApps
-						)
-						
-						backupOptionToggle(
-							icon: "square.and.arrow.down.fill",
-							iconColor: .orange,
-							title: .localized("Imported Apps"),
-							description: .localized("Apps imported from files or other sources"),
-							isOn: $options.includeImportedApps
-						)
-						
-						backupOptionToggle(
-							icon: "globe.fill",
-							iconColor: .purple,
-							title: .localized("Sources"),
-							description: .localized("Your configured app sources and repositories"),
-							isOn: $options.includeSources
-						)
-
-						backupOptionToggle(
-							icon: "puzzlepiece.extension.fill",
-							iconColor: .cyan,
-							title: .localized("Default Frameworks"),
-							description: .localized("Your automatically injected frameworks (.dylib, .deb)"),
-							isOn: $options.includeDefaultFrameworks
-						)
-					}
-					.padding(.horizontal, 20)
-					
-					// Warning notice
-					if options.includeSignedApps || options.includeImportedApps {
-						HStack(alignment: .top, spacing: 12) {
-							Image(systemName: "exclamationmark.triangle.fill")
-								.font(.system(size: 20))
-								.foregroundStyle(.orange)
-							
-							VStack(alignment: .leading, spacing: 4) {
-								Text(.localized("Large Backup Size"))
-									.font(.headline)
-									.foregroundStyle(.primary)
-								
-								Text(.localized("If you include Signed and Imported Apps, this backup will be large."))
-									.font(.subheadline)
-									.foregroundStyle(.secondary)
-									.fixedSize(horizontal: false, vertical: true)
-							}
-						}
-						.padding(16)
-						.background(Color.orange.opacity(0.1))
-						.cornerRadius(12)
-						.padding(.horizontal, 20)
-					}
-					
-					// Create button
-					Button {
-						onConfirm()
-					} label: {
-						HStack {
-							Image(systemName: "checkmark.circle.fill")
-							Text(.localized("Create Backup"))
-								.font(.headline)
-						}
-						.frame(maxWidth: .infinity)
-						.padding(.vertical, 16)
-						.background(
-							LinearGradient(
-								gradient: Gradient(colors: [.blue, .purple]),
-								startPoint: .leading,
-								endPoint: .trailing
-							)
-						)
-						.foregroundStyle(.white)
-						.cornerRadius(12)
-					}
-					.padding(.horizontal, 20)
-					.padding(.top, 8)
-					
-					// Cancel button
-					Button {
-						dismiss()
-					} label: {
-						Text(.localized("Cancel"))
-							.font(.subheadline)
-							.foregroundStyle(.secondary)
-					}
-					.padding(.bottom, 20)
-				}
-			}
-			.navigationTitle(.localized("Backup Options"))
-			.navigationBarTitleDisplayMode(.inline)
-			.toolbar {
-				ToolbarItem(placement: .navigationBarTrailing) {
-					Button {
-						dismiss()
-					} label: {
-						Image(systemName: "xmark.circle.fill")
-							.font(.system(size: 20))
-							.foregroundStyle(.secondary)
-					}
-				}
-			}
-		}
-		.presentationDetents([.large])
-		.presentationDragIndicator(.visible)
-	}
-	
-	@ViewBuilder
-	private func backupOptionToggle(
-		icon: String,
-		iconColor: Color,
-		title: LocalizedStringKey,
-		description: LocalizedStringKey,
-		isOn: Binding<Bool>
-	) -> some View {
-		Button {
-			isOn.wrappedValue.toggle()
-			HapticsManager.shared.softImpact()
-		} label: {
-			HStack(alignment: .top, spacing: 12) {
-				ZStack {
-					Circle()
-						.fill(iconColor.opacity(0.15))
-						.frame(width: 44, height: 44)
-					
-					Image(systemName: icon)
-						.font(.system(size: 20, weight: .semibold))
-						.foregroundStyle(iconColor)
-				}
-				
-				VStack(alignment: .leading, spacing: 4) {
-					Text(title)
-						.font(.headline)
-						.foregroundStyle(.primary)
-					
-					Text(description)
-						.font(.subheadline)
-						.foregroundStyle(.secondary)
-						.fixedSize(horizontal: false, vertical: true)
-				}
-				
-				Spacer()
-				
-				Image(systemName: isOn.wrappedValue ? "checkmark.circle.fill" : "circle")
-					.font(.system(size: 24))
-					.foregroundStyle(isOn.wrappedValue ? .blue : .gray.opacity(0.3))
-			}
-			.padding(16)
-			.background(Color(uiColor: .secondarySystemGroupedBackground))
-			.cornerRadius(12)
-		}
-		.buttonStyle(.plain)
-	}
-}
-
-// MARK: - RestoreLoadingOverlay
-struct RestoreLoadingOverlay: View {
-	let progress: Double
-	@State private var rotation: Double = 0
-	
-	var body: some View {
-		ZStack {
-			// Blurred background
-			Color.black.opacity(0.5)
-				.ignoresSafeArea()
-			
-			// Card
-			VStack(spacing: 24) {
-				// Animated icon
-				ZStack {
-					Circle()
-						.fill(
-							LinearGradient(
-								colors: [Color.green.opacity(0.2), Color.green.opacity(0.1)],
-								startPoint: .topLeading,
-								endPoint: .bottomTrailing
-							)
-						)
-						.frame(width: 100, height: 100)
-					
-					Circle()
-						.stroke(Color.green.opacity(0.3), lineWidth: 4)
-						.frame(width: 100, height: 100)
-						.rotationEffect(.degrees(rotation))
-					
-					Image(systemName: "arrow.down.circle.fill")
-						.font(.system(size: 50))
-						.foregroundStyle(
-							LinearGradient(
-								colors: [.green, .green.opacity(0.7)],
-								startPoint: .topLeading,
-								endPoint: .bottomTrailing
-							)
-						)
-				}
-				
-				// Text
-				VStack(spacing: 8) {
-					Text("Restoring Backup")
-						.font(.title2.bold())
-						.foregroundStyle(.white)
-					
-					Text("Please wait while we restore your data...")
-						.font(.subheadline)
-						.foregroundStyle(.white.opacity(0.8))
-						.multilineTextAlignment(.center)
-				}
-				
-				// Progress bar
-				VStack(spacing: 8) {
-					GeometryReader { geometry in
-						ZStack(alignment: .leading) {
-							// Background
-							RoundedRectangle(cornerRadius: 8)
-								.fill(Color.white.opacity(0.2))
-								.frame(height: 8)
-							
-							// Progress
-							RoundedRectangle(cornerRadius: 8)
-								.fill(
-									LinearGradient(
-										colors: [.green, .green.opacity(0.8)],
-										startPoint: .leading,
-										endPoint: .trailing
-									)
-								)
-								.frame(width: geometry.size.width * progress, height: 8)
-								.animation(.easeInOut(duration: 0.3), value: progress)
-						}
-					}
-					.frame(height: 8)
-					
-					Text("\(Int(progress * 100))%")
-						.font(.caption)
-						.foregroundStyle(.white.opacity(0.8))
-				}
-			}
-			.padding(32)
-			.background(
-				RoundedRectangle(cornerRadius: 20)
-					.fill(Color(uiColor: .systemBackground))
-					.opacity(0.95)
-			)
-			.padding(.horizontal, 40)
-			.shadow(color: .black.opacity(0.3), radius: 20, x: 0, y: 10)
-		}
-		.onAppear {
-			withAnimation(.linear(duration: 1.5).repeatForever(autoreverses: false)) {
-				rotation = 360
-			}
-		}
-	}
 }
